refactor: extract HSM definitions from state-machine.ts to hsm-definitions.ts
